extract PositionCard
